Tests/LibWeb/Text/expected/hit_testing/stacking-context-with-inline-content.txt

node: #text
text: LMAO

indexInNode: 1
